css: don't hard code search entry height
When commit 961fdd861fac46ef0ecc67bb8be14c0ce828e612 landed
text in the overview search entry field started getting clipped.

This is because the default font size changed but the entry is a
hard coded pixel height.

This commit drops the hard coding, so the entry can automatically
size itself to the proper height.

https://bugzilla.gnome.org/show_bug.cgi?id=641537