build: Modify maintainer compiler flags values
We use the micro version for distinguishing released tarballs and Git
builds; the maintainer compiler flags should be enabled for the latter,
and not just for unstable cycles, since it makes sense to have extra
warning flags even on stable cycles.

We also want to allow people to turn on -Werror on demand, so let's add
a third option to --enable-maintainer-flags.