Remove minimize and maximize buttons from title bar
The minimize and maximize buttons on the titlebar have lost much of their
utility in GNOME 3: the minimize functionality is largely superceded by
the ability to switch windows in the overview and conveniently use
multiple workspaces, while the maximize functionality is available by
drag to the top gesture or double clicking on the title bar.

Removing these buttons significantly cleans up the default appearance
and avoids exposing a half-done implementation of minimization.

http://mail.gnome.org/archives/gnome-shell-list/2011-February/msg00192.html
https://bugzilla.gnome.org/show_bug.cgi?id=604237